App for the lesson 5: adding functionality to DAO

+++ module_L4/src/main/java/org/example/services/EntityService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface EntityService<T> {
 
-    T getById(int id);
+    T getById(Long id);
 
     void add(T t);
 
@@ -12,6 +12,6 @@ public interface EntityService<T> {
 
     void update(T t);
 
-    void remove(int id);
+    void remove(Long id);
 
 }
